requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.1
dask==2023.2.0
distlib==0.3.6
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
Faker==17.0.0
filelock==3.9.0
flake8==6.0.0
fsspec==2023.1.0
gunicorn==20.1.0
identify==2.5.18
idna==3.4
kombu==5.2.4
locket==1.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
partd==1.3.0
pathspec==0.11.0
platformdirs==3.0.0
pre-commit==3.0.4
prompt-toolkit==3.0.37
psycopg2==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toolz==0.12.0
urllib3==1.26.14
vine==5.0.0
virtualenv==20.19.0
wcwidth==0.2.6
